workshops/deploy-aws/buyer/requirements.txt
strands-agents[bedrock]>=1.0.0
payments-py>=1.3.0
fastapi>=0.100.0
uvicorn>=0.20.0
httpx>=0.24.0
python-dotenv>=1.0.0
boto3>=1.28.0
sse-starlette>=1.0.0
a2a-sdk>=0.3.0
